refactor(sandbox): move legacy registry migration into doctor
